refactor: Global code refactoring
